Add InfraRabbitMQ project with RabbitMQ integration
A new project named "InfraRabbitMQ" has been added to the solution `Project-microservice.sln`. This project is defined in the file `InfraRabbitMQ\InfraRabbitMQ.csproj`.

The build configurations for the new "InfraRabbitMQ" project have been added to the solution, including Debug and Release settings for Any CPU.

The `InfraRabbitMQ.csproj` file targets the .NET 8.0 framework, with ImplicitUsings and Nullable enabled. It includes package references for `Microsoft.Extensions.Options` version 8.0.2 and `RabbitMQ.Client` version 6.8.1.

A new class `RabbitMQSettings` has been added in the `InfraRabbitMQ` namespace, containing a constant string `Settings` and a property `HostName`.

A new class `ProjectSyncHandler` has been added in the `InfraRabbitMQ.Handler.DataSync` namespace. This class uses `IOptions<RabbitMQSettings>` to retrieve RabbitMQ settings, initializes a `ConnectionFactory` with the `HostName` from the settings, and throws an exception if the settings are not found.