Add Kubernetes update job and update solution items
A new job named "Kubernetes update" has been added to the
`Prodcution_Pipeline.yml` file. This job is dependent on the
"Deploy" job and runs on the latest Ubuntu environment. It
includes steps to set the Kubernetes context using the
`Azure/k8s-set-context@v1` action with a kubeconfig method and
a secret for the kubeconfig. Additionally, it includes a step
to restart the Kubernetes deployment named `project-app` in
the `deanban` namespace using the `kubectl rollout restart`
command.

The `Prodcution_Pipeline.yml` file has been added to the
solution items in the `Project-microservice.sln` file. This is
indicated by the addition of the line
`.github\workflows\Prodcution_Pipeline.yml =
.github\workflows\Prodcution_Pipeline.yml` under the `Solution
Items` project section.

Similarly, the `On_Pull_Request.yml` file has also been added
to the solution items in the `Project-microservice.sln` file.